Updating names and moving empty mocks to separate file
